fix: :recycle: seperate db utils into seperate modules to fix db generate command
`npm run db:generate` command broke because all utils where being called instead of only needed modules (`date`)
